Fix delegate route
